Switch address infrastructure to interfaces.
The previous system had one master address struct, however, now all
of the user level functions take an Address interface. The only
issue with this is that it causes problems when unmarshalling an
address into json which is why there is still a default StreetAddr
struct that is easily created from an Address and unmarshalled into
json.