Added the structs and interfaces to impliment the changed Item system.
The Item system is a system that better conforms to dominos menu and ordering system.
The current implimentaion uses one object to represent three objects that are
inherently different. The new Item system (built around the Item interface) has three
different structs that represent a product category, a product, and a product that houses
data sent in an order.